chore: renamed is to readable
